Dynamic Turret subclass
