+++ Tower Defence/Assets/Scripts/Turrets/DynamicTurret.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9d0ed3edd8fcf9e41a875f0183a285d7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
